fgcedfdsv

+++ lockers/models.py
@@ -4,5 +4,5 @@ from django.db import models
 # Create your models here.
 
 class Cart(models.Model):
-    number = models.IntegerField()
+    number = models.BigIntegerField()
     owner = models.CharField(max_length=1000)
